refactor(codec-sensor): extract operator strategies + LoupeRenderer (closes #326)
Per the AI-shape audit (docs/research/2026-05-13-ai-shape-audit.md,
merged via #328), packages/codec-sensor/src/render.ts (447 lines)
carried two canonical AI-code smells:

1. expandOperator() was a 79-line switch over 7 operator types with
   each branch carrying time-arithmetic and noise-gen inline.
2. expandPatchGrammar() called expandOperator recursively while managing
   timeOriginFrame offsets — patch scheduling and operator semantics
   were tangled.
3. renderLoupeDashboard() had a 3-level fallback chain (loupe.py search
   → loupe_cli on PATH → static-HTML fallback) with near-identical
   try/catch blocks.

This refactor splits along all three seams:

- packages/codec-sensor/src/operators/ (new, 8 files):
  - oscillator.ts, noise.ts (with whiteNoise + pinkNoise helpers),
    drift.ts, pulse.ts, step.ts, ecg-template.ts, patch-grammar.ts
  - index.ts: shared OperatorContext type + dispatchOperator(operator,
    ctx) — a tight switch that routes to the strategy implementation
  - Each operator file owns its own expansion logic and is
    independently testable / replaceable
  - patch-grammar.ts uses dispatchOperator recursively for inner
    operators (small same-folder cycle that ES modules handle fine
    since dispatch isn't called at module-load time)

- packages/codec-sensor/src/loupe-renderer.ts (new):
  - Encapsulates the 3-level fallback chain behind a single
    renderLoupeDashboard(csvPath, htmlPath, spec) API
  - SensorRenderPath outcome string preserved for manifest stability
  - Configurable search paths (defaults to original 4-level lookup
    order) for future testability

- packages/codec-sensor/src/render.ts (shrunk 447 → 206L):
  - Now a thin orchestrator: builds samples via dispatchOperator,
    writes JSON/CSV sidecars, runs the Loupe renderer for the HTML
    companion
  - makeEcgSignal / makeGyroSignal / makeTemperatureSignal helpers
    preserved with same signatures (they call expandSensorAlgorithm
    which is also preserved, just internally using dispatchOperator
    instead of the old switch)

What did NOT change:

- Sample byte output. The refactor moves code around but the actual
  per-frame math is character-identical to the pre-refactor switch.
  All 6 sensor goldens preserve byte-for-byte.
- Public API. renderSignalBundle, makeEcgSignal, makeGyroSignal,
  makeTemperatureSignal, SensorSample, SensorRenderPath all keep
  the same signatures.
- patchGrammar semantics. The recursion-seam invariant
  (single-patch full-duration matches flat operator list byte-for-byte)
  is preserved — patch-grammar.ts passes the same arguments to
  dispatchOperator that the inline expandOperator received.
- Loupe fallback paths. The 4 candidate paths for loupe.py and the
  fallback static-HTML body are byte-identical.

Validation:

- pnpm --filter @wittgenstein/codec-sensor typecheck — clean.
- pnpm --filter @wittgenstein/codec-sensor test — all 18 tests pass
  including all 6 golden tests (the patchGrammar byte-stability
  fixtures are the recursion-seam regression baseline).
- pnpm --filter @wittgenstein/codec-sensor lint — 0 errors, 0 warnings.

render.ts shrinks 447 → 206 lines (54% reduction; under the audit's
"under 250 lines" target). Total LOC across the 10 files (1 root + 1
loupe + 8 operators) is slightly higher than the original (one-line
file headers and per-operator JSDoc cost real space), which is the
expected cost of the seam.

Refs:
- Closes #326.
- Audit source: docs/research/2026-05-13-ai-shape-audit.md (merged via
  #328).
- Sibling refactor PRs: #337 (landscape renderer extraction, just
  merged), #327 (HyperFrames composition + ProcessRunner, queued).
- Related sensor work: #284 patchGrammar measurement gate — unaffected
  by this refactor since patchGrammar semantics are preserved.
